+++ jspm.config.js
@@ -50,7 +50,7 @@ SystemJS.config({
     "plugin-typescript": "github:frankwallis/plugin-typescript@5.0.9",
     "readline": "github:jspm/nodelibs-readline@0.2.0-alpha",
     "text": "github:systemjs/plugin-text@0.0.7",
-    "angular2-carbonldp": "npm:angular2-carbonldp@0.3.0-rc.7",
+    "angular2-carbonldp": "npm:angular2-carbonldp@0.3.0",
     "assert": "github:jspm/nodelibs-assert@0.2.0-alpha",
     "buffer": "github:jspm/nodelibs-buffer@0.2.0-alpha",
     "carbonldp": "npm:carbonldp@0.36.0",
@@ -362,7 +362,174 @@ SystemJS.config({
         "typescript": "npm:typescript@2.0.0"
       }
     },
-    "npm:angular2-carbonldp@0.3.0-rc.7": {
+    "npm:request@2.74.0": {
+      "map": {
+        "aws-sign2": "npm:aws-sign2@0.6.0",
+        "forever-agent": "npm:forever-agent@0.6.1",
+        "combined-stream": "npm:combined-stream@1.0.5",
+        "form-data": "npm:form-data@1.0.0-rc4",
+        "har-validator": "npm:har-validator@2.0.6",
+        "json-stringify-safe": "npm:json-stringify-safe@5.0.1",
+        "hawk": "npm:hawk@3.1.3",
+        "is-typedarray": "npm:is-typedarray@1.0.0",
+        "mime-types": "npm:mime-types@2.1.11",
+        "bl": "npm:bl@1.1.2",
+        "isstream": "npm:isstream@0.1.2",
+        "stringstream": "npm:stringstream@0.0.5",
+        "qs": "npm:qs@6.2.1",
+        "tunnel-agent": "npm:tunnel-agent@0.4.3",
+        "node-uuid": "npm:node-uuid@1.4.7",
+        "oauth-sign": "npm:oauth-sign@0.8.2",
+        "extend": "npm:extend@3.0.0",
+        "caseless": "npm:caseless@0.11.0",
+        "http-signature": "npm:http-signature@1.1.1",
+        "aws4": "npm:aws4@1.4.1",
+        "tough-cookie": "npm:tough-cookie@2.3.1"
+      }
+    },
+    "npm:form-data@1.0.0-rc4": {
+      "map": {
+        "combined-stream": "npm:combined-stream@1.0.5",
+        "mime-types": "npm:mime-types@2.1.11",
+        "async": "npm:async@1.5.2"
+      }
+    },
+    "npm:bl@1.1.2": {
+      "map": {
+        "readable-stream": "npm:readable-stream@2.0.6"
+      }
+    },
+    "npm:har-validator@2.0.6": {
+      "map": {
+        "pinkie-promise": "npm:pinkie-promise@2.0.1",
+        "chalk": "npm:chalk@1.1.3",
+        "is-my-json-valid": "npm:is-my-json-valid@2.13.1",
+        "commander": "npm:commander@2.9.0"
+      }
+    },
+    "npm:mime-types@2.1.11": {
+      "map": {
+        "mime-db": "npm:mime-db@1.23.0"
+      }
+    },
+    "npm:hawk@3.1.3": {
+      "map": {
+        "hoek": "npm:hoek@2.16.3",
+        "cryptiles": "npm:cryptiles@2.0.5",
+        "sntp": "npm:sntp@1.0.9",
+        "boom": "npm:boom@2.10.1"
+      }
+    },
+    "npm:combined-stream@1.0.5": {
+      "map": {
+        "delayed-stream": "npm:delayed-stream@1.0.0"
+      }
+    },
+    "npm:http-signature@1.1.1": {
+      "map": {
+        "assert-plus": "npm:assert-plus@0.2.0",
+        "sshpk": "npm:sshpk@1.9.2",
+        "jsprim": "npm:jsprim@1.3.0"
+      }
+    },
+    "npm:sntp@1.0.9": {
+      "map": {
+        "hoek": "npm:hoek@2.16.3"
+      }
+    },
+    "npm:readable-stream@2.0.6": {
+      "map": {
+        "inherits": "npm:inherits@2.0.1",
+        "isarray": "npm:isarray@1.0.0",
+        "process-nextick-args": "npm:process-nextick-args@1.0.7",
+        "core-util-is": "npm:core-util-is@1.0.2",
+        "string_decoder": "npm:string_decoder@0.10.31",
+        "util-deprecate": "npm:util-deprecate@1.0.2"
+      }
+    },
+    "npm:cryptiles@2.0.5": {
+      "map": {
+        "boom": "npm:boom@2.10.1"
+      }
+    },
+    "npm:sshpk@1.9.2": {
+      "map": {
+        "assert-plus": "npm:assert-plus@1.0.0",
+        "dashdash": "npm:dashdash@1.14.0",
+        "getpass": "npm:getpass@0.1.6",
+        "asn1": "npm:asn1@0.2.3"
+      }
+    },
+    "npm:pinkie-promise@2.0.1": {
+      "map": {
+        "pinkie": "npm:pinkie@2.0.4"
+      }
+    },
+    "npm:is-my-json-valid@2.13.1": {
+      "map": {
+        "xtend": "npm:xtend@4.0.1",
+        "generate-function": "npm:generate-function@2.0.0",
+        "jsonpointer": "npm:jsonpointer@2.0.0",
+        "generate-object-property": "npm:generate-object-property@1.2.0"
+      }
+    },
+    "npm:boom@2.10.1": {
+      "map": {
+        "hoek": "npm:hoek@2.16.3"
+      }
+    },
+    "npm:chalk@1.1.3": {
+      "map": {
+        "escape-string-regexp": "npm:escape-string-regexp@1.0.5",
+        "has-ansi": "npm:has-ansi@2.0.0",
+        "strip-ansi": "npm:strip-ansi@3.0.1",
+        "ansi-styles": "npm:ansi-styles@2.2.1",
+        "supports-color": "npm:supports-color@2.0.0"
+      }
+    },
+    "npm:jsprim@1.3.0": {
+      "map": {
+        "extsprintf": "npm:extsprintf@1.0.2",
+        "json-schema": "npm:json-schema@0.2.2",
+        "verror": "npm:verror@1.3.6"
+      }
+    },
+    "npm:dashdash@1.14.0": {
+      "map": {
+        "assert-plus": "npm:assert-plus@1.0.0"
+      }
+    },
+    "npm:commander@2.9.0": {
+      "map": {
+        "graceful-readlink": "npm:graceful-readlink@1.0.1"
+      }
+    },
+    "npm:getpass@0.1.6": {
+      "map": {
+        "assert-plus": "npm:assert-plus@1.0.0"
+      }
+    },
+    "npm:generate-object-property@1.2.0": {
+      "map": {
+        "is-property": "npm:is-property@1.0.2"
+      }
+    },
+    "npm:verror@1.3.6": {
+      "map": {
+        "extsprintf": "npm:extsprintf@1.0.2"
+      }
+    },
+    "npm:strip-ansi@3.0.1": {
+      "map": {
+        "ansi-regex": "npm:ansi-regex@2.0.0"
+      }
+    },
+    "npm:has-ansi@2.0.0": {
+      "map": {
+        "ansi-regex": "npm:ansi-regex@2.0.0"
+      }
+    },
+    "npm:angular2-carbonldp@0.3.0": {
       "map": {
         "js-cookie": "npm:js-cookie@2.1.2"
       }